Move FakeState into its own file
